Current status

+++ Snippets/TarGapsFiles.py
@@ -16,7 +16,7 @@ def get_runs(datasets = None):
     dataset_selection = ''
 
     if datasets is not None:
-        dataset_selection = "AND u.dataset_id IN (%s)" % ','.join([str(d) for d in datasets])
+        dataset_selection = "u.dataset_id IN (%s) AND" % ','.join([str(d) for d in datasets])
 
     sql = """SELECT 
     run_id, u.dataset_id, path, name, urlpath_id, u.queue_id
@@ -26,9 +26,8 @@ FROM
     i3filter.run r ON r.queue_id = u.queue_id
         AND r.dataset_id = u.dataset_id
 WHERE
-    u.dataset_id < 1874
         %s
-        AND name LIKE '%%gaps.txt'""" % dataset_selection
+        name LIKE '%%gaps.txt'""" % dataset_selection
 
     dbresult = dbs4.MySQL().fetchall(sql, UseDict = True)
 
@@ -47,53 +46,81 @@ WHERE
 
     return result
 
-def skip_run(run_id, gaps_files):
+def skip_run(run_id, gaps_files, deleteonly = False):
     # Check if tar file already exists
     tar_file = os.path.join(os.path.dirname(gaps_files[-1]['path']), "Run00%s_GapsTxt.tar" % run_id)
 
-    print "Check if %s exists" % tar_file
+    exists = os.path.isfile(tar_file)
 
-    return os.path.isfile(tar_file)
+    print "Check if %s exists (%s)" % (tar_file, exists)
 
-def tar_gaps(run_id, gaps_files):
+    if exists:
+        # If we want to create tar files, then we know that we can skip this run. If we want to delete gaps file, we need further checks
+        if not deleteonly:
+            return True
+    elif deleteonly:
+        # If the tar-ed file doesn't exist and we only want to delete gaps-files, we want to skip this run
+        return True
+
+    print "Check if gaps files exists"
+
+    for file in gaps_files:
+        if not os.path.isfile(file['path']):
+            print "******** File %s does not exist" % file
+            return True
+
+    return False
+
+def tar_gaps(run_id, gaps_files, dryrun):
     dest = os.path.join(os.path.dirname(gaps_files[-1]['path']), "Run00%s_GapsTxt.tar" % run_id)
 
     print "Tar file: %s" % dest
 
-    with tarfile.open(dest, 'w') as tar:
-        for file in gaps_files:
-            tar.add(file['path'], arcname = os.path.basename(file['path']))
-
+    if not dryrun:
+        with tarfile.open(dest, 'w') as tar:
+            for file in gaps_files:
+                tar.add(file['path'], arcname = os.path.basename(file['path']))
+    
     return dest
 
-def fix_db(dataset_id, tar_file, gaps_files):
-    tar_path = 'file:' + os.path.dirname(tar_file)
-    tar_name = os.path.basename(tar_file)
-
-    size = os.path.getsize(tar_file)
-    md5sum = FileTools(tar_file).md5sum()
+def fix_db(dataset_id, tar_file, gaps_files, deleteonly, dryrun):
+    if not deleteonly:
+        tar_path = 'file:' + os.path.dirname(tar_file)
+        tar_name = os.path.basename(tar_file)
+    
+        size = os.path.getsize(tar_file)
+        md5sum = FileTools(tar_file).md5sum()
 
     # The tar gaps files get the queue id of the last sub run
     queue_id = gaps_files[-1]['queue_id']
 
     sql = "UPDATE urlpath SET transferstate = 'DELETED' WHERE urlpath_id IN (%s)" % ','.join(str(f['uid']) for f in gaps_files)
+
+    if deleteonly:
+        sql = "DELETE FROM urlpath WHERE urlpath_id IN (%s)" % ','.join(str(f['uid']) for f in gaps_files)
+
     #print "UPDATE SQL statement: %s" % sql
     
-    dbs4.MySQL().execute(sql)
+    if not dryrun:
+        dbs4.MySQL().execute(sql)
 
-    sql = "INSERT INTO urlpath (name, path, type, dataset_id, queue_id, md5sum, size, transferstate) VALUES ('%s', '%s', '%s', %s, %s, '%s', %s, '%s')" % (
-        tar_name, tar_path, 'PERMANENT', dataset_id, queue_id, md5sum, size, 'WAITING'
-    )
+    if not deleteonly:
+        sql = "INSERT INTO urlpath (name, path, type, dataset_id, queue_id, md5sum, size, transferstate) VALUES ('%s', '%s', '%s', %s, %s, '%s', %s, '%s')" % (
+            tar_name, tar_path, 'PERMANENT', dataset_id, queue_id, md5sum, size, 'WAITING'
+        )
+    
+        #print "INSERT SQL statement: %s" % sql
+    
+        if not dryrun:
+            dbs4.MySQL().execute(sql)
 
-    #print "INSERT SQL statement: %s" % sql
-
-    dbs4.MySQL().execute(sql)
-
-def move_gaps_files(gaps_files):
+def move_gaps_files(gaps_files, dest, dryrun):
     for file in gaps_files:
-        d = os.path.join('/data/user/i3filter/tmp/gaps_files', os.path.basename(file['path']))
+        d = os.path.join(dest, os.path.basename(file['path']))
         print "move %s to %s" % (file['path'], d)
-        shutil.move(file['path'], d)
+
+        if not dryrun:
+            shutil.move(file['path'], d)
 
 def show_datasets(runs):
     def get_info(info, dataset_id):
@@ -119,7 +146,11 @@ def show_datasets(runs):
 
 parser = argparse.ArgumentParser()
 parser.add_argument('--datasets', nargs = "*", help = "Only specific datasets", type = int)
+parser.add_argument('--datasetinfo', action = 'store_true', default = False, help = "Print only affected datasets")
+parser.add_argument('--dryrun', action = 'store_true', default = False, help = "Run the script w/o doing something")
+parser.add_argument('--deleteonly', action = 'store_true', default = False, help = "Do not create tar files. Do only delete *_gaps.txt files. Note: The gaps files are only deleted if the tar file already exists")
 parser.add_argument('-n', nargs = "?", help = "Number of runs that should be processed", type = int)
+parser.add_argument('--gapsdest', nargs = "?", help = "Destination of the moved gaps files. Default is /data/user/i3filter/tmp/gaps_files", type = str, default = '/data/user/i3filter/tmp/gaps_files')
 args = parser.parse_args()
 
 max_runs = args.n
@@ -128,6 +159,9 @@ runs = get_runs(datasets = args.datasets)
 
 # Show list of affected datasets
 show_datasets(runs)
+
+if args.datasetinfo:
+    exit()
 
 counter = 0
 
@@ -146,13 +180,21 @@ for dataset_id, runs in runs.iteritems():
         print '================================================================'
         print "\n"
 
-        if skip_run(run_id, gaps_files):
-            print "Skip run because tar-ed gaps file already exist"
+        if skip_run(run_id, gaps_files, deleteonly = args.deleteonly):
+            if args.deleteonly:
+                print "Skip run because no tar file exists"
+            else:
+                print "Skip run because tar-ed gaps file already exist"
+
             continue
 
-        tar_file = tar_gaps(run_id, gaps_files)
-        fix_db(dataset_id, tar_file, gaps_files)
-        move_gaps_files(gaps_files)
+        tar_file = None
+        if not args.deleteonly:
+            tar_file = tar_gaps(run_id, gaps_files, dryrun = args.dryrun)
+        
+        fix_db(dataset_id, tar_file, gaps_files, dryrun = args.dryrun, deleteonly = args.deleteonly)
+
+        move_gaps_files(gaps_files, dest = args.gapsdest, dryrun = args.dryrun)
 
         print "\n"
